The things that needed to be fixed.
1) Checking from one point past the origin. Ln 150

2) Checking of the current and forward slope must be done at the
beginning of the loop Ln 171 to 174.

3) If the current point is less than the origin then set it to null and
keep counting the similar slopes this allows us to skip the section of
the array with similar slopes but not with the lowest origin. Similar
reasoning when we compare the highest point to the origin. Ln 184 and Ln
200.

4) Add the streakCounter to the current index if it was greater than 3
and only draw the line if the highest point was not null. Ln 210 and Ln
222.